+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * Fixed a bug that caused navigation information not to be loaded when switching maps
  * Prevent from segfault on failing SignalReference identification when loading OpenDrive files
  * Added vehicle doors to the recorder
+ * Adjusted vehicle BoundingBox when the vehicle opens the doors.
  * Added functions to get actor' components transform
  * Added posibility to Digital Twins to work with local files (osm and xodr)
  * Enable proper material merging for Building in Digital Twins